Added prepare script and extend it to show which modules have uncommitted changes
