--- scripts/deploy-steno
@@ -1,23 +1,0 @@
-#!/bin/bash
-
-build_dir=$(pwd)/dist
-public_url=/next/ui-concepts/
-
-# Deploy ui concepts to dev.macrostrat.org/next/ui-concepts/
-pushd concept-apps
-parcel build \
-  --dist-dir $build_dir \
-  --public-url $public_url \
-  apps/index.html
-popd
-
-pushd storybook
-
-# Deploy storybook to dev.macrostrat.org/next/ui-concepts/stories/
-npx build-storybook --modern --output-dir $build_dir/stories
-
-popd
-# Upload the files
-rsync -azv --delete $build_dir/ steno:/data/projects/macrostrat$public_url
-
-
